Update patch set 2
Patch Set 2:

I want to note that, due to the timestamp issue that we raised to Vijay, this POM is currently insufficient for CICD. The global docker jjb is not currently used in ONAP DCAE. I believe we are waiting for Vijay to clarify whether we need the timestamps or not. 
If we do, this pom will have to grow (by a lot).

Patch-set: 2