Update patch set 2
Patch Set 2:

The functionality that was in the Dockerfiles is moved into the pom.xml.  This puts all the build process in one place and also takes care of things like only having to update one place for version changes.  The Dockerfiles are removed as part of this so nobody comes along and tries to update them, and is then mystified that their changes don't seem to have any effect.

Patch-set: 2