Update patch set 2
Patch Set 2:

The docker/README.md file should be deleted as it is now inaccurate, obsolete, and irrelevant.  ONAP uses Kubernetes and Helm, not manual deployment of individual images on isolated Docker hosts, and the builds use Maven to compile and build images, not manual invocation of docker build.

Patch-set: 2