Update patch set 2
Patch Set 2:

At the time this was submitted, no version of the Acumos Adapter had been released, therefore there were no official versions out there that might conflict and therefore no version bump was required.  If an official version of the Acumos Adapter has since been released, with a conflicting version, it is possible that the delay in merging this, may have rendered a version bump necessary.

Patch-set: 2