Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> However, since the CSIT team wants images always be tagged "latest" and the release team wants images to always be tagged with the build timestamp to the second, the version of the image in the pom.xml doesn't actually affect anything.

1.0.0 was released 2020-03-04.

When this is merged, presumably we will get a timestamped image in the snapshot tree on nexus3. We'll create a release file that points at that timestamped image, and supply an appropriate tag (!== "1.0.0") as part of the release file.

Patch-set: 2
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0