Update patch set 1
Patch Set 1:

Why are you asking for bumping of the version number when the image with the currently version number is unreleased?  What will break if the version number is not changed, and, if nothing, what benefit is there from changing it?  Is there a reason to only fix one thing per version number?

Patch-set: 1