Update patch set 3
Patch Set 3:

Some background for the change - I wanted to make it more clear that the bpgen is supposed to be used as library, so I separated logic for commandline reading to separate module. Also that way it will be easier to use in runtime-api (no issues that the bpgen has a class annotated with @SpringBootApplication etc)

Last patchset (3) I reworked the BlueprintJarComparatorTest in order to:
 - make it a little easier to maintain
 - make files cleanup happen before and after test run
 - make it more system independent (change of "\\" as path separator to File.separator)
 - remove SpringContext startup from the test (not needed)

I also compared the jar from this code to jar built from the origin/master, the tests passed.

Patch-set: 3