Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Why is this split necessary? The executable was generated under onap module build earlier - separating it would add overhead to match versions across pom.

Currently there is no build for lib jar for runtime-api.

I think that the split is quite "natural", and shows that the base bpgen is a library used elsewhere. It's also easier to use in runtime api that way (I'm not sure how to "disable" the main method when importing it in runtime api, it tampers with tests etc., maybe using multiple executions for spring boot plugin would be a way). If you prefer to keep it as single submodule I can look into it a bit more.

Regarding versions - use of mvn versions:set -DnewVersion=<version> ensures that the versions match.

Patch-set: 5