Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (7 comments)
> 
> minor stuff that needs to be resolved

Thank you for your support. Will make the changes and commit again.
Regarding the pyyaml module, the requirements.txt file is a dictionary that is extracted from the model's metadata.json file. And Acumos is not including pyyaml, only cv, dill, acumos etc. I am not sure if the DCAE model runner included pyyaml before as this is a recent error with the current release of DCAE model runner. Currently, when the pod starts, the DCAE model runner throws an error as pyyaml module not found. I can only say that it uses pyyaml without installing it exclusively or probably assumes that the environment has pyyaml installed.

Patch-set: 4